ai-engine/requirements_pi.txt
ultralytics==8.1.0
opencv-python-headless
pandas
numpy
joblib
supabase
python-dotenv
